[refactor/#120] 스케쥴, 기록 리스트 공통 컴포넌트화
