--- app/src/main/java/com/bongtu/baekseo/presentation/schedule/component/ScheduleListContent.kt
@@ -1,210 +1,0 @@
-package com.bongtu.baekseo.presentation.schedule.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.asPaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.navigationBars
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.bongtu.baekseo.R.string.record_card_list_month
-import com.bongtu.baekseo.R.string.record_card_list_year
-import com.bongtu.baekseo.core.common.type.ScheduleCardType
-import com.bongtu.baekseo.core.designsystem.component.card.BongBaekScheduleCard
-import com.bongtu.baekseo.core.designsystem.theme.BongBaekTheme
-import com.bongtu.baekseo.core.util.OnBottomReached
-import com.bongtu.baekseo.data.model.event.ScheduleEvent
-import com.bongtu.baekseo.presentation.schedule.model.ScheduleYearMonthEventItem
-import com.bongtu.baekseo.presentation.schedule.model.toYearMonthEventItemList
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
-
-// TODO: 기록하기와 공통 컴포넌트화 할 것
-@Composable
-fun ScheduleListContent(
-    scheduleEventList: ImmutableList<ScheduleEvent>,
-    onCardClick: (String) -> Unit,
-    lazyListState: LazyListState,
-    updatePage: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val yearMonthEventItems = scheduleEventList.toYearMonthEventItemList()
-    val navigationBarsPadding =
-        WindowInsets.navigationBars.asPaddingValues().calculateBottomPadding()
-
-    val hasUserScrolled = remember(lazyListState) {
-        derivedStateOf {
-            lazyListState.firstVisibleItemIndex > 0 || lazyListState.isScrollInProgress
-        }
-    }
-
-    LazyColumn(
-        modifier = modifier
-            .fillMaxSize(),
-        state = lazyListState,
-        contentPadding = PaddingValues(
-            start = 20.dp,
-            top = 20.dp,
-            end = 20.dp,
-            bottom = 20.dp + navigationBarsPadding,
-        ),
-    ) {
-        itemsIndexed(
-            items = yearMonthEventItems,
-        ) { index, item ->
-            val previousItem = yearMonthEventItems.getOrNull(index - 1)
-
-            val topPadding = remember(item, previousItem) {
-                when (item) {
-                    is ScheduleYearMonthEventItem.YearHeader -> if (previousItem == null) 0.dp else 40.dp
-                    is ScheduleYearMonthEventItem.MonthHeader -> 30.dp
-                    is ScheduleYearMonthEventItem.Event -> {
-                        if (previousItem is ScheduleYearMonthEventItem.Event) 10.dp else 20.dp
-                    }
-                }
-            }
-
-            when (item) {
-                is ScheduleYearMonthEventItem.YearHeader -> {
-                    Text(
-                        text = stringResource(record_card_list_year, item.year),
-                        color = BongBaekTheme.colors.white,
-                        style = BongBaekTheme.typography.headBold24,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = topPadding),
-                    )
-                }
-
-                is ScheduleYearMonthEventItem.MonthHeader -> {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 4.dp)
-                            .padding(top = topPadding),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Text(
-                            text = stringResource(record_card_list_month, item.month),
-                            color = BongBaekTheme.colors.white,
-                            style = BongBaekTheme.typography.titleSemiBold16,
-                            modifier = Modifier.padding(end = 12.dp),
-                        )
-                        HorizontalDivider(
-                            thickness = 1.dp,
-                            color = BongBaekTheme.colors.gray750,
-                        )
-                    }
-                }
-
-                is ScheduleYearMonthEventItem.Event -> {
-                    with(item.event) {
-                        BongBaekScheduleCard(
-                            scheduleCardType = ScheduleCardType.SCHEDULE,
-                            hostName = hostName,
-                            hostNickname = hostNickname,
-                            eventCategory = eventCategory,
-                            relationship = relationship,
-                            cost = cost,
-                            eventDate = eventDate,
-                            onCardClick = { onCardClick(eventId) },
-                            modifier = Modifier.padding(top = topPadding),
-                        )
-                    }
-                }
-            }
-        }
-    }
-
-    if (hasUserScrolled.value) {
-        lazyListState.OnBottomReached(
-            buffer = 3,
-            onLoadMore = updatePage,
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun ScheduleListContentPreview() {
-    BongBaekTheme {
-        ScheduleListContent(
-            modifier = Modifier.background(color = BongBaekTheme.colors.gray900),
-            scheduleEventList = persistentListOf(
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.02.11 (목)",
-                ),
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.09.11 (목)",
-                ),
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.08.11 (목)",
-                ),
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.07.11 (목)",
-                ),
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.06.11 (목)",
-                ),
-                ScheduleEvent(
-                    eventId = "1",
-                    hostName = "공승준",
-                    hostNickname = "초록승준",
-                    eventCategory = "결혼식",
-                    relationship = "친구",
-                    cost = 10000,
-                    eventDate = "2025.05.11 (목)",
-                ),
-            ),
-            onCardClick = {},
-            lazyListState = rememberLazyListState(),
-            updatePage = {},
-        )
-    }
-}
